project/requirements.txt
numpy
midiutil
pandas
matplotlib
pillow
requests
tensorflow
keras
opencv-python
scipy
librosa
music21
flask
flask-cors
google-cloud-storage
vamp
python-dotenv
pysynth